Partial rewrite of out Midi file output routines. Doesn't compile yet!
More consistent use of types 'byte' and 'dword' to (hopefully) increase portability.
Replaced all calls to MacOS FSWrite() with new function WriteRawBytes() that uses stdio.h.
Split header writing code out into a new function. There is some very similar code in NewTrack() and CloseMIDIFile() that could probably be refactored.

+++ source/BP2/-BP2.h
@@ -1664,7 +1664,7 @@ typedef struct s_chunck ChunkPointer;
 typedef int (*IntProcPtr)(int);
 typedef int (*Int2ProcPtr)(short,int);
 
-typedef unsigned long dword;
+typedef uint32_t dword;
 typedef unsigned char byte;
 
 typedef int tokenbyte;
